+++ playwright/tests/authorSearch.spec.ts
@@ -3,7 +3,7 @@ import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
 
 Given("I want to do an author search", async function (this: CustomWorld) {
-  return await this.page.goto("https://digital-research-books-beta.nypl.org/");
+  return await this.page.goto(`${this.parameters.appUrl}`);
 });
 When("I change the dropdown to author", async function (this: CustomWorld) {
   const dropdown = await this.page.$("[aria-label='Select a search category']");