+++ playwright/tests/keywordSearch.spec.ts
@@ -5,9 +5,7 @@ import { CustomWorld } from "../support/setup";
 Given(
   "I am on the DRB home page for searching",
   async function (this: CustomWorld) {
-    return await this.page.goto(
-      "https://digital-research-books-beta.nypl.org/"
-    );
+    return await this.page.goto(`${this.parameters.appUrl}`);
   }
 );