+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add features and tests for keyword search
 - add features and tests for author search
 - add features and tests for title search
+- add features and tests for subjects search
 - Upgrade to NYPL Design System version 1.1.0
 - Upgrade to React 17.0.2
 - Upgrade to Next 12.2.5